# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server used for POS TUI.",
        "title": "Swagger POS TUI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-git clone TODO: CHANGE"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
...
        "/order_lines": {
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Create a new orderLine with the given input data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderLines"
                ],
                "summary": "Create a new orderLine",
                "parameters": [
                    {
                        "description": "Create orderLine object",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateOrderLine"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created orderLine",
                        "schema": {
                            "$ref": "#/definitions/models.OrderLine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/order_lines/{id}": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Get details of a orderLine by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderLines"
                ],
                "summary": "Find a orderLine by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OrderLine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved orderLine",
                        "schema": {
                            "$ref": "#/definitions/models.OrderLine"
                        }
                    },
                    "404": {
                        "description": "OrderLine not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Update the orderLine details for the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderLines"
                ],
                "summary": "Update a orderLine by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OrderLine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update orderLine object",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateOrderLine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated orderLine",
                        "schema": {
                            "$ref": "#/definitions/models.OrderLine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "orderLine not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Delete the orderLine with the given ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderLines"
                ],
                "summary": "Delete a orderLine by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OrderLine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted orderLine",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "orderLine not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateOrderLine": {
            "type": "object",
            "required": [
                "price",
                "product_id",
                "quantity",
                "total",
                "vat"
            ],
            "properties": {
                "price": {
                    "description": "In Cents, with VAT",
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "description": "decimal.NewFromString(\"136.02\")",
                    "type": "number"
                },
                "total": {
                    "description": "In Cents",
                    "type": "integer"
                },
                "vat": {
                    "description": "(ex: 2100 for 21.00%)",
                    "type": "integer"
                }
            }
        },
...
        "models.OrderLine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "description": "In Cents, with VAT",
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "description": "decimal.NewFromString(\"136.02\")",
                    "type": "number"
                },
                "total": {
                    "description": "In Cents, with VAT",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "vat": {
                    "description": "(ex: 2100 for 21.00%)",
                    "type": "integer"
                }
            }
        },
...
        "models.UpdateOrderLine": {
            "type": "object",
            "properties": {
                "price": {
                    "description": "In Cents, with VAT",
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "description": "decimal.NewFromString(\"136.02\")",
                    "type": "number"
                },
                "total": {
                    "description": "In Cents, with VAT",
                    "type": "integer"
                },
                "vat": {
                    "description": "(ex: 2100 for 21.00%)",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}